Organize library components
